Replace CompBinClean counter with IsHashIntervalTick call
Multiple bins would end up with the same counter value when reloading the game, or when multiple bins were spawned on the same tick. This could end up in lag spikes when all bins performed their operations on the same tick.

On top of that, the code was iterating over all spawned filth in home area even if the bin was full.

This PR should fix that by:
- Removing the counter
- Checking if the operation should be performed using `IsHashIntervalTick`
- Checking beforehand if the bin is not full

On top of that, `CanAccept` method no longer checks if the bin is full as:
- The method calling it now does it before calling it
- It would still return `false` if it was full