Most of XML, some C# done
I've done most of XML needed for the update. I've included some of the required C#, but more needs to be done.

The code I've made for modular couch was not included, as it'll be included in the framework. Some other features are unfinished.

I've rearranged the graphic files, trying moving them to separate folders to avoid clutter.