Consolidated computer JobDef/JoyGiverDefs into single one
The computer usage JobDef and JoyGiverDefs were separate for both industrial and modern computer. I've decided it would be the best to consolidate them, similarly to watching TV. This will ensure that building both types of PCs won't result in that joy type being more common. It should also allow for adding more computer variants easier in addons.

The XP gain rate and skill are defined in `ExtendedSitFacingJoyDataExtension`, as using the JobDef's fields for it didn't allow us to customize those on a building-by-building basis.